Move NodesManager from search to ingest ORM
