+++ nucliadb/nucliadb/search/lifecycle.py
@@ -21,11 +21,11 @@ import logging
 import sys
 
 from nucliadb.ingest import logger as ingest_logger
+from nucliadb.ingest.orm.nodes_manager import NodesManager
 from nucliadb.ingest.utils import get_driver  # type: ignore
 from nucliadb.ingest.utils import start_ingest, stop_ingest
 from nucliadb.search import SERVICE_NAME, logger
 from nucliadb.search.chitchat import start_chitchat, stop_chitchat
-from nucliadb.search.nodes import NodesManager
 from nucliadb.search.predict import PredictEngine
 from nucliadb_telemetry.utils import clean_telemetry, get_telemetry, init_telemetry
 from nucliadb_utils.settings import nuclia_settings, running_settings
